pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "silvaengine-utility"
version = "0.0.6"
description = "High-performance utility library with JSON optimization and AWS integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "bibow", email = "bibow@silvaengine.org"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    "requests>=2.28.0",
    "python-dateutil>=2.8.0",
    "sqlalchemy>=1.4.0"
]

[project.optional-dependencies]
performance = [
    "orjson>=3.8.0",
    "pendulum>=2.1.0"
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]
all = [
    "orjson>=3.8.0",
    "pendulum>=2.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/silvaengine/silvaengine-utility"
Repository = "https://github.com/silvaengine/silvaengine-utility"
Documentation = "https://silvaengine-utility.readthedocs.io"
"Bug Tracker" = "https://github.com/silvaengine/silvaengine-utility/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["silvaengine_utility*"]

[tool.setuptools.package-data]
silvaengine_utility = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance benchmarks"
]

[tool.coverage.run]
source = ["silvaengine_utility"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]